+++ test/tests/server.node.js
@@ -1,0 +1,7 @@
+var assert = require('assert');
+var nodeserverClass = require('../../nodeserver');
+
+
+module.exports = exports = {
+	
+};